requirements.txt
Flask==2.3.5
Flask-SocketIO==6.7.2
requests==2.31.0
eventlet==0.33.3
discord.py==2.3.1
